oboe: move OpenSL ES output code to its own class
