+++ src/opensles/AudioInputStreamOpenSLES.h
@@ -32,6 +32,21 @@ public:
     explicit AudioInputStreamOpenSLES(const AudioStreamBuilder &builder);
 
     virtual ~AudioInputStreamOpenSLES();
+
+    Result open() override;
+    Result close() override;
+
+    Result requestStart() override;
+    Result requestPause() override;
+    Result requestFlush() override;
+    Result requestStop() override;
+
+    Result waitForStateChange(StreamState currentState,
+                              StreamState *nextState,
+                              int64_t timeoutNanoseconds) override;
+
+    int chanCountToChanMask(int chanCount) override;
+
 };
 
 } // namespace oboe
